Add support for Nemesis-compressed tile data.
This was awkward to implement: mdcomp (AKA FW-KENSC) relies on Boost,
so I have to rely on an older version of that lacks this requirement.
However, older versions of mdcomp are bugged and do not compile.
Luckily, I forked and fixed an old version of mdcomp a long time ago
for MainMemory's FW-KENSC shell extension. This forks works great, so
I'm using it here.