Improve sprite importing.
When importing sprites, SonMapEd seems to be using a method that
involves determining the closest RGB value in the palette. This is
naive, since RGB does not depict colours in a 'uniformly linear'
sense: a red of 128 is not half as red as a red of 256. Likewise, a
red of 256 is not as bright as a blue of 256. This means that it's
possible for two colours to have a similar RGB value while looking
nothing alike, while two colours with wildly different RGB values can
look almost identical.

To avoid this, my editor now uses LAB instead of RGB. This appears to
produces more-pleasing results.